dj_shop/requirements.txt
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
Babel==2.6.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cssselect2==0.4.1
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distro==1.5.0
Django==3.1.4
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-parler==2.2
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-rosetta==0.9.5
docker==4.4.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gpg===1.13.1-unknown
grpcio==1.16.1
html5lib==1.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
ldap3==2.4.1
localflavor==1.8
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
nemo-emblems==4.6.0
netifaces==0.10.4
nose==1.3.7
oauthlib==3.1.0
onboard==1.4.1
PAM==0.4.2
paramiko==2.6.0
passlib==1.7.2
pbr==5.5.1
pexpect==4.6.0
phonenumbers==8.12.15
Pillow==7.0.0
ply==3.11
polib==1.1.0
protobuf==3.6.1
protonvpn-cli==2.2.6
psutil==5.5.1
psycopg2==2.8.5
psycopg2-binary==2.8.4
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparted==3.11.2
Pyphen==0.10.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.17.3
PySocks==1.6.8
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
python-xapp==2.0.1
python-xlib==0.23
pythondialog==3.5.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.15.0
soupsieve==1.9.5
speaklater==1.4
SQLAlchemy==1.3.12
sqlparse==0.4.1
sshtunnel==0.1.4
systemd-python==234
testresources==2.0.1
texttable==1.6.3
tinycss2==1.0.2
toml==0.10.2
torbrowser-launcher==0.3.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
wadllib==1.3.3
WeasyPrint==52.2
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
wrapt==1.12.1
WTForms==2.2.1
xkit==0.0.0
zeroconf==0.24.4
